Movido el ConectandoDialog a una clase publica
